requirements.txt
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
numpy==1.22.1
pandas==1.4.0
pdfminer.six==20211012
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tabulate==0.8.9
